+++ common/src/com/giancola/battleship/gamelogic/GameLogic.kt
@@ -16,9 +16,9 @@ abstract class GameLogic {
     protected var playersData: MutableMap<PlayerId, PlayerData> = mutableMapOf()
 
     protected abstract fun notifyGameStarted()
-    protected abstract fun notifyCombatStarted(playerTurn: PlayerId)
+    protected abstract fun notifyCombatStarted(playerTurn: PlayerId, playerNames: Map<PlayerId, String>)
     protected abstract fun notifyGameFinished(winner: PlayerId)
-    protected abstract fun notifyShot(playerId: PlayerId, gridX: Int, gridY: Int, shotResult: ShotResult?)
+    protected abstract fun notifyShot(shooter: PlayerId, gridX: Int, gridY: Int, shotResult: ShotResult?)
 
     private var nextPlayerId: PlayerId? = PlayerId.Player1
 
@@ -76,6 +76,6 @@ abstract class GameLogic {
         playerTurn = PlayerId.values().random()
         gameRunning = true
 
-        notifyCombatStarted(playerTurn)
+        notifyCombatStarted(playerTurn, playersData.keys.associateWith { playersData[it]!!.name })
     }
 }